+++ src/Method/RemoveMyProfilePhoto.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phptg\BotApi\Method;
+
+use Phptg\BotApi\MethodInterface;
+use Phptg\BotApi\ParseResult\ValueProcessor\TrueValue;
+use Phptg\BotApi\Transport\HttpMethod;
+
+/**
+ * @see https://core.telegram.org/bots/api#removemyprofilephoto
+ *
+ * @template-implements MethodInterface<true>
+ *
+ * @api
+ */
+final readonly class RemoveMyProfilePhoto implements MethodInterface
+{
+    public function getHttpMethod(): HttpMethod
+    {
+        return HttpMethod::POST;
+    }
+
+    public function getApiMethod(): string
+    {
+        return 'removeMyProfilePhoto';
+    }
+
+    public function getData(): array
+    {
+        return [];
+    }
+
+    public function getResultType(): TrueValue
+    {
+        return new TrueValue();
+    }
+}